:recycle: Compose Article  모듈 분리
