+++ Unit1/Pathway3/TaskCompleted/app/src/main/res/values/strings.xml
@@ -3,9 +3,6 @@
     <!-- Task Completed Screen -->
     <string name="all_task_completed">All tasks completed</string>
     <string name="nice_work">Nice work!</string>
-    <string name="text_title">Jetpack Compose tutorial</string>
-    <string name="text_introduce">Jetpack Compose is a modern toolkit for building native Android UI. Compose simplifies and accelerates UI development on Android with less code, powerful tools, and intuitive Kotlin APIs.\n</string>
-    <string name="text_content">In this tutorial, you build a simple UI component with declarative functions. You call Compose functions to say what elements you want and the Compose compiler does the rest. Compose is built around Composable functions. These functions let you define your app\'s UI programmatically because they let you describe how it should look and provide data dependencies, rather than focus on the process of the UI\'s construction, such as initializing an element and then attaching it to a parent. To create a Composable function, you add the @Composable annotation to the function name.</string>
     <string name="text_task_state">All tasks completed</string>
     <string name="text_notification">Nice work!</string>
